requirements.txt
streamlit==1.39.0
ultralytics==8.3.63
opencv-python==4.10.0.84
Pillow==10.4.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
matplotlib>=3.8.0
